refactor : post컴포넌트 위치 이동
