+++ client/src/components/index.tsx
@@ -2,4 +2,4 @@ export { default as Advertisement } from '@/components/Advertisement';
 export { default as Layout } from '@/components/Layout';
 export { default as MenuView } from '@/components/MenuView';
 export { default as Navigator } from '@/components/Navigator';
-export { default as Post } from '@/components/Post';
+export { Post, PostDetail } from '@/components/Post';
